Take the tabPrototype to its own module out of tabState.js
