+++ js/tabState/tab.js
@@ -1,0 +1,133 @@
+var tabPrototype = {
+  add: function (tab, index) {
+    // make sure the tab exists before we create it
+    if (!tab) {
+      var tab = {}
+    }
+
+    var tabId = String(tab.id || Math.round(Math.random() * 100000000000000000)) // you can pass an id that will be used, or a random one will be generated.
+
+    var newTab = {
+      url: tab.url || '',
+      title: tab.title || '',
+      id: tabId,
+      lastActivity: tab.lastActivity || Date.now(),
+      secure: tab.secure,
+      private: tab.private || false,
+      readerable: tab.readerable || false,
+      backgroundColor: tab.backgroundColor,
+      foregroundColor: tab.foregroundColor,
+      selected: tab.selected || false
+    }
+
+    if (index) {
+      this.splice(index, 0, newTab)
+    } else {
+      this.push(newTab)
+    }
+
+    return tabId
+  },
+  update: function (id, data) {
+    if (!this.has(id)) {
+      throw new ReferenceError('Attempted to update a tab that does not exist.')
+    }
+    var index = -1
+    for (var i = 0; i < this.length; i++) {
+      if (this[i].id === id) {
+        index = i
+      }
+    }
+    for (var key in data) {
+      if (data[key] === undefined) {
+        throw new ReferenceError('Key ' + key + ' is undefined.')
+      }
+      this[index][key] = data[key]
+    }
+  },
+  destroy: function (id) {
+    for (var i = 0; i < this.length; i++) {
+      if (this[i].id === id) {
+        tasks.getTaskContainingTab(id).tabHistory.push(this[i])
+        this.splice(i, 1)
+        return i
+      }
+    }
+    return false
+  },
+  destroyAll: function () {
+    // this = [] doesn't work, so set the length of the array to 0 to remove all of the itemss
+    this.length = 0
+  },
+  get: function (id) {
+    if (!id) { // no id provided, return an array of all tabs
+      // it is important to deep-copy the tab objects when returning them. Otherwise, the original tab objects get modified when the returned tabs are modified (such as when processing a url).
+      var tabsToReturn = []
+      for (var i = 0; i < this.length; i++) {
+        tabsToReturn.push(JSON.parse(JSON.stringify(this[i])))
+      }
+      return tabsToReturn
+    }
+    for (var i = 0; i < this.length; i++) {
+      if (this[i].id === id) {
+        return JSON.parse(JSON.stringify(this[i]))
+      }
+    }
+    return undefined
+  },
+  has: function (id) {
+    for (var i = 0; i < this.length; i++) {
+      if (this[i].id === id) {
+        return true
+      }
+    }
+    return false
+  },
+  getIndex: function (id) {
+    for (var i = 0; i < this.length; i++) {
+      if (this[i].id === id) {
+        return i
+      }
+    }
+    return -1
+  },
+  getSelected: function () {
+    for (var i = 0; i < this.length; i++) {
+      if (this[i].selected) {
+        return this[i].id
+      }
+    }
+    return null
+  },
+  getAtIndex: function (index) {
+    return this[index] || undefined
+  },
+  setSelected: function (id) {
+    if (!this.has(id)) {
+      throw new ReferenceError('Attempted to select a tab that does not exist.')
+    }
+    for (var i = 0; i < this.length; i++) {
+      if (this[i].id === id) {
+        this[i].selected = true
+      } else {
+        this[i].selected = false
+      }
+    }
+  },
+  count: function () {
+    return this.length
+  },
+  isEmpty: function () {
+    if (!this || this.length === 0) {
+      return true
+    }
+
+    if (this.length === 1 && (!this[0].url || this[0].url === 'about:blank')) {
+      return true
+    }
+
+    return false
+  }
+}
+
+module.exports = tabPrototype
